+++ threedollar-api-admin-service/src/e2e-test/kotlin/com/depromeet/threedollar/api/adminservice/config/advice/ControllerExceptionAdviceTest.kt
@@ -1,4 +1,4 @@
-package com.depromeet.threedollar.api.adminservice.controller.advice
+package com.depromeet.threedollar.api.adminservice.config.advice
 
 import org.junit.jupiter.api.Test
 import org.springframework.http.converter.HttpMessageNotReadableException
