+++ threedollar-api-user-service/src/integration-test/java/com/depromeet/threedollar/api/userservice/service/auth/AppleAuthServiceTest.java
@@ -99,6 +99,7 @@ class AppleAuthServiceTest extends IntegrationTest {
                 .socialType(SOCIAL_TYPE)
                 .build();
 
+            // when
             authService.signUp(request);
 
             // then
