+++ threedollar-common/src/main/java/com/depromeet/threedollar/common/constants/VersionConstants.kt
@@ -2,6 +2,6 @@ package com.depromeet.threedollar.common.constants
 
 object VersionConstants {
 
-    const val VERSION: String = "3.0.4"
+    const val VERSION: String = "3.0.5"
 
 }
